requirements.txt
asgiref
asn1crypto
blinker
certifi
cffi
chardet
charset-normalizer
click
colorama
coreapi
coreschema
cryptography
defusedxml
dj-database-url
Django
django-allauth
django-cors-headers
django-heroku
django-storages
datetime
idna
feedparser
httplib2
django-filter
django-multiselectfield
django-rest-auth
django-restframework
djangorestframework
djangorestframework-simplejwt
djoser
lazr.restfulclient
lazr.uri
gunicorn
Jinja2
lxml
MarkupSafe
Pillow
PyJWT
pytz
python3-openid
requests
requests-oauthlib
requests-unixsocket
SecretStorage
six
social-auth-app-django
social-auth-core
sqlparse
soupsieve
uritemplate
urllib3
whitenoise
python-dotenv
beautifulsoup
requests

